@@ -7,7 +7,7 @@ import fr.insee.rmes.modelSwagger.dataset.Label;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.annotation.Generated;
+import jakarta.annotation.Generated;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
