@@ -8,8 +8,8 @@ import fr.insee.rmes.model.datasets.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.annotation.Generated;
-import javax.validation.Valid;
+import jakarta.annotation.Generated;
+import jakarta.validation.Valid;
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.List;
