@@ -16,7 +16,7 @@ import org.json.JSONObject;
 import org.springframework.stereotype.Service;
 import com.fasterxml.jackson.databind.JsonNode;
 
-import javax.xml.soap.SAAJResult;
+import jakarta.xml.soap.SAAJResult;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
